+++ src/main/java/org/folio/sidecar/exception/KeycloakUnhandledAuthorizationException.java
@@ -1,0 +1,14 @@
+package org.folio.sidecar.exception;
+
+import lombok.Getter;
+
+@Getter
+public class KeycloakUnhandledAuthorizationException extends RuntimeException {
+
+  private final int statusCode;
+
+  public KeycloakUnhandledAuthorizationException(int statusCode) {
+    super("Authorization service error");
+    this.statusCode = statusCode;
+  }
+}